Add flag --mock-port / -m to specify port for Mock
When running pact-go in containers we want to expose minimal number of
ports. Currently it randomly selects _any_ port. This makes it
impossible to expose ports in a container.

This change will add a CLI option to specify a port to deploy the
mocking service on. This way we know exactly which port to expose.