feat(verification): Expose verification output
BREAKING CHANGE: PactClient#VerifyProvider now returns a
types.ProviderVerifierResponse as well as an error. An error
means that the client was unable to run the verification code
successfully. The ProviderVerifierResponse is a direct mapping
of the json output returned by the pact-provider-verifier binary.
Iterate over the Examples structs to determine if they passed.

NOTE: Verbose output for the types.VerifyRequest is not compatible
with the json format since it logs to stdout, causing the json to
be unparsable.