feat(matching): deprecate matching JSON formatted objects
- Removes ability for user to supply JSON document
in matching DSL.
- Warning displayed to console if used

BREAKING CHANGE

JSON formatted objects such as `{"foo": "bar"}` are no
longer supported in the matching DSL. Marshalling these
to the Mock Service correctly is problematic, especially
when combining with matchers and other nested JSON
formatted objects.

Type information is lost re-assembling into a workable
map[string]interface{} which meant that, for example, a "100"
is converted to 100 losing key information.

If we detect a JSON formatted object in the DSL, we now log
a warning indicating this potential problem.

See #73 for
background.